backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.5",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.9",
    "langchain-nvidia-ai-endpoints>=1.0.3",
    "langgraph>=1.0.8",
    "numpy>=2.4.2",
    "pathlib>=1.0.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
